refactor(credential data source): duplicate method removed
