+++ internal/utils/markdown_helpers_test.go
@@ -46,3 +46,30 @@ func TestNewIntMarkdownList(t *testing.T) {
 
 	assert.Equal(t, []int64{1}, got.ToInt64())
 }
+
+func TestStringTypeArrayToMarkdown(t *testing.T) {
+	type underlyingString string
+
+	enumValues := []underlyingString{
+		"TEST_ONE",
+		"TEST_TWO",
+	}
+
+	want := "\n  - *TEST_ONE*\n  - *TEST_TWO*\n"
+	got := StringTypeArrayToMarkdown(enumValues)
+	assert.Equal(t, want, got)
+}
+
+func ExampleStringTypeArrayToMarkdown() {
+	type underlyingString string
+
+	enumValues := []underlyingString{
+		"TEST_ONE",
+		"TEST_TWO",
+	}
+
+	markdown := StringTypeArrayToMarkdown(enumValues)
+
+	fmt.Println(markdown)
+	// Output "\n  - *TEST_ONE*\n  - *TEST_TWO*\n"
+}
